% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_clust.R
\name{em_clust}
\alias{em_clust}
\title{EM Clustering}
\usage{
EM(x, k)
}
\arguments{
\item{x}{numeric matrix of data, or an object that can be coerced to such a matrix
(such as a numeric vector or a data frame with all numeric columns).}

\item{k}{the number of initial means}
}
\value{
A list containing cluster means, clustering vector, and covariance.
}
\description{
Preform EM clustering on a dataset.
}
